fix: corregir violaciones de calidad y eliminar código duplicado en módulo de firma
- Extraer lógica de notificaciones duplicada a SignatureFlowNotificationService compartido entre CreateSignatureFlowUseCase y ProcessFlowParticipantActionUseCase
- Eliminar método muerto notifyResponsibleOnRejection de ValidateSignatureCodeUseCase y sus parámetros de constructor sin uso (inAppNotificationRepository, emailService)
- Cambiar ValidationError a ServerError en errores de configuración de Twilio en InitiateSignatureUseCase
- Corregir typo en mensaje SMS: "Tú código" → "Tu código"
- Agregar validación de participantes en CreateSignatureFlowUseCase (requiere userId o externalEmail)
- Traducir mensajes de error de LinkUserToColaboratorUseCase al español
- Actualizar dependency-container con el nuevo servicio de notificaciones
