+++ src/domains/signature-flow/use-cases/create-signature-flow.use-case.ts
@@ -4,20 +4,12 @@ import { type SignatureFlowRepository } from '../repositories/signature-flow.rep
 import { type SignatureFlowParticipantRepository } from '../repositories/signature-flow-participant.repository';
 import { DocumentRepository } from '@domains/document/repositories/document.repository';
 import { DocumentHistoryRepository } from '@domains/document/repositories/document-history.repository';
-import { UserRepository } from '@domains/user/repositories/user.repository';
-import { InAppNotificationRepository } from '@domains/notification/repositories/in-app-notification.repository';
-import { InAppNotification } from '@domains/notification/entities/in-app-notification.entity';
 import { DocumentAction, DocumentStatus } from '@domains/document/value-objects/document-enums';
 import {
-  SignatureFlowOrderType,
   SignatureFlowParticipantRole,
   SignatureFlowStatus,
 } from '../value-objects/signature-flow-enums';
-import { EmailService } from '@shared/infrastructure/email/email-service.interface';
-import {
-  buildFrontendUrl,
-  buildPrimactaNotificationEmail,
-} from '@shared/infrastructure/email/templates/primacta-notification-email.template';
+import { SignatureFlowNotificationService } from '../services/signature-flow-notification.service';
 import { NotFoundError, ValidationError } from '@shared/domain/errors';
 
 export interface CreateSignatureFlowInput {
@@ -39,14 +31,18 @@ export class CreateSignatureFlowUseCase {
     private readonly participantRepository: SignatureFlowParticipantRepository,
     private readonly documentRepository: DocumentRepository,
     private readonly documentHistoryRepository: DocumentHistoryRepository,
-    private readonly userRepository: UserRepository,
-    private readonly inAppNotificationRepository: InAppNotificationRepository,
-    private readonly emailService: EmailService,
+    private readonly notificationService: SignatureFlowNotificationService,
   ) {}
 
   async execute(input: CreateSignatureFlowInput): Promise<SignatureFlow> {
     if (!input.participants || input.participants.length === 0) {
       throw new ValidationError('El flujo debe tener al menos un participante');
+    }
+
+    for (const p of input.participants) {
+      if (!p.userId && !p.externalEmail) {
+        throw new ValidationError('Cada participante debe tener un usuario asignado o un correo externo');
+      }
     }
 
     const document = await this.documentRepository.findById(input.documentId);
@@ -55,7 +51,7 @@ export class CreateSignatureFlowUseCase {
     }
 
     const now = new Date();
-    const hasValidators = input.participants.some((participant) => participant.role === SignatureFlowParticipantRole.VALIDATOR);
+    const hasValidators = input.participants.some((p) => p.role === SignatureFlowParticipantRole.VALIDATOR);
 
     const flowProps: SignatureFlowProps = {
       documentId: input.documentId,
@@ -78,8 +74,8 @@ export class CreateSignatureFlowUseCase {
         role: p.role,
         order: p.order ?? null,
       };
-      const savedParticipant = await this.participantRepository.save(SignatureFlowParticipant.create(participantProps));
-      savedParticipants.push(savedParticipant);
+      const saved = await this.participantRepository.save(SignatureFlowParticipant.create(participantProps));
+      savedParticipants.push(saved);
     }
 
     document.signatureFlowId = flow.id;
@@ -101,77 +97,11 @@ export class CreateSignatureFlowUseCase {
       comment: 'Documento enviado al flujo de firma',
     });
 
-    await this.notifyParticipantsForCurrentStep(document.id, document.name, flow.orderType, savedParticipants);
+    const validators = savedParticipants.filter((p) => p.role === SignatureFlowParticipantRole.VALIDATOR);
+    const signers = savedParticipants.filter((p) => p.role === SignatureFlowParticipantRole.SIGNER);
+    const toNotify = validators.length > 0 ? validators : signers;
+    await this.notificationService.notifyParticipantsForCurrentStep(toNotify, document.id, document.name, flow.orderType);
 
     return flow;
-  }
-
-  private async notifyParticipantsForCurrentStep(
-    documentId: string,
-    documentName: string,
-    orderType: SignatureFlowOrderType,
-    participants: SignatureFlowParticipant[],
-  ): Promise<void> {
-    const validators = participants.filter((participant) => participant.role === SignatureFlowParticipantRole.VALIDATOR);
-    const signers = participants.filter((participant) => participant.role === SignatureFlowParticipantRole.SIGNER);
-    const participantsToNotify = validators.length > 0
-      ? this.pickParticipantsToNotify(orderType, validators)
-      : this.pickParticipantsToNotify(orderType, signers);
-
-    for (const participant of participantsToNotify) {
-      if (!participant.userId) continue;
-
-      const isValidator = participant.role === SignatureFlowParticipantRole.VALIDATOR;
-      const title = isValidator ? 'Documento pendiente de revision' : 'Documento pendiente de firma';
-      const message = isValidator
-        ? `Tienes un documento pendiente de revisar: ${documentName}`
-        : `Tienes un documento pendiente de firmar: ${documentName}`;
-
-      await this.inAppNotificationRepository.save(new InAppNotification({
-        userId: participant.userId,
-        title,
-        message,
-        entityType: 'document',
-        entityId: documentId,
-      }));
-
-      const user = await this.userRepository.findById(participant.userId);
-      if (!user?.email) continue;
-
-      const actionUrl = buildFrontendUrl(`/signature-flows?documentId=${encodeURIComponent(documentId)}`);
-      const html = buildPrimactaNotificationEmail({
-        title,
-        recipientName: user.firstName,
-        message,
-        actionLabel: 'Ir a pendientes',
-        actionUrl,
-      });
-
-      await this.emailService.send({
-        to: user.email.toString(),
-        subject: title,
-        text: actionUrl ? `${message}\n\nIr a pendientes: ${actionUrl}` : message,
-        html,
-      });
-    }
-  }
-
-  private pickParticipantsToNotify(
-    orderType: SignatureFlowOrderType,
-    participants: SignatureFlowParticipant[],
-  ): SignatureFlowParticipant[] {
-    if (orderType !== SignatureFlowOrderType.SEQUENTIAL) {
-      return participants;
-    }
-
-    const orderedParticipants = participants.filter((participant) => participant.order !== null);
-    if (orderedParticipants.length === 0) {
-      return participants;
-    }
-
-    const firstOrder = Math.min(...orderedParticipants.map((participant) => participant.order as number));
-    const firstBatch = participants.filter((participant) => participant.order === firstOrder);
-
-    return firstBatch.length > 0 ? firstBatch : participants;
   }
 }
